project-bolt-sb1-5evnf2ne/project/requirements.txt
Django==5.0.2
Pillow==10.1.0
mutagen==1.47.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-dotenv==1.0.0
openai==1.3.0